pyproject.toml

[project]
name = "yolo-fastapi-app"
version = "0.1.0"
description = "Object detection FastAPI application"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pillow>=8.0.0",
    "opencv-python-headless>=4.5.0",
    "numpy>=1.20.0",
    "python-multipart>=0.0.5"
]
